+++ lib/yard-sorbet/handlers/sig_handler.rb
@@ -48,8 +48,8 @@ module YARDSorbet
         return if ATTR_NODE_TYPES.include?(method_node.type)
 
         sibling = NodeUtils.sibling_node(node)
-        sibling[0][0].each do |param|
-          param_name = param[0][0]
+        sibling.dig(0, 0).each do |param|
+          param_name = param.dig(0, 0)
           types = SigToYARD.convert(param.last)
           TagUtils.upsert_tag(docstring, 'param', types, param_name)
         end